+++ misc/illustration_horizontal_fluxes.png
